backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
anthropic==0.5.0
openai==1.3.0
gradio-client==0.5.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
langchain-anthropic==0.3.0
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.3.0
langchain-community-anthropic==0.3.0
